Add missing files
I realised that `wget` had not downloaded me all files because some were not linked to from within the wiki so `wget` didn't know about them. The ones mainly affected are user files.

We want to preserve links from previous programme pages, so I have run `wget` on the programme pages dating back to when the programme linked out to user pages.

We're not interested in changes in existing files as they are mostly changes of timestamp; we're only interested in new files (i.e. missing from the initial crawl).