Do not return 300 Multiple Files if no file
Prior to this change, if a close match to the filename couldn't be found, Apache would return a 300 and display around 50 links that were allegedly quite similar to the one you'd been looking for. This wasn't helpful and looks bad. I'd prefer a 404 for a file not found.

I've set this in the mediawiki directory rather than on the root directory because this does have useful features for the main site, for example allowing incorrect case in the URLs and allowing one character different; see for example https://www.spaconference.org/spa2017/LOCATIONO.html.