Add a redirect for URLs with `&` something
Some of the URLs have extra query parameters that no longer make sense and range from having no effect, through giving an unstyled page, to returning a 404 when the page is actually there (because the title is not exactly the same).

This rule looks for query strings that have a second parameter and discards it.

Rewrite Conditions are executed in order so this has to be before the one that looks for any characters at all.

Test data, all of which now works:

````
http://www.spaconference.org/cgi-bin/wiki.pl?browse=JuhaPekkaTolvanen&type=bio2
http://www.spaconference.org/cgi-bin/wiki.pl?browse=DavidHarvey
http://www.spaconference.org/mediawiki/index.php?title=User:RachelDavies
http://www.spaconference.org/scripts/people.php?username=NathanielPryce&useskin=bio
http://www.spaconference.org/mediawiki/index.php?title=User:RachelDavies&useskin=bio
http://www.spaconference.org/mediawiki/index.php?title=User:DickGabriel&type=bio
````