Redirect `/mediawiki/` to wiki main page
Before this change, `/mediawiki` was a 301 to `/mediawiki/` because of CheckSpelling, and `/mediawiki/` was a 403 because Apache thinks you are trying to view the contents of that folder and there is no `index.html` present.

However, clearly someone with a link to this is looking for the `mediawiki` main page.

This redirects a request with no path (beyond `mediawiki`) and no query string to the main page.