Delete opensearch
Separating this into lines didn't seem to make a difference, there were still errors (on line 1!). But we don't need this. The search on the wiki site doesn't work so this is not being used. Deleting it so far does not appear to have caused any issues, so removing it also from the repo.